move utils to a central point
